# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SANC Logistics API",
    "description": "시온식품(주) 물류부 주문 관리 시스템 API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "tags": [
    { "name": "System", "description": "API health / info" },
    { "name": "Auth", "description": "로그인" },
    { "name": "Members", "description": "회원 가입 및 아이디 중복 확인" },
    { "name": "Orders", "description": "주문 조회 / 생성" }
  ],
  "paths": {
    "/api": {
      "get": {
        "tags": ["System"],
        "summary": "API status",
        "responses": {
          "200": {
            "description": "API is online",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "online" },
                    "message": { "type": "string" },
                    "version": { "type": "string" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "로그인",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string", "example": "tenorseon" },
                  "password": { "type": "string", "example": "your-password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "username": { "type": "string" },
                        "name": { "type": "string" },
                        "role": { "type": "string", "enum": ["admin", "member"] }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "입력값 누락" },
          "401": { "description": "아이디 또는 비밀번호 불일치" },
          "500": { "description": "서버 오류" }
        }
      }
    },
    "/api/members/check-username": {
      "get": {
        "tags": ["Members"],
        "summary": "아이디 중복 확인",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9]{4,20}$",
              "example": "leesh01"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "중복 확인 결과",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "available": { "type": "boolean" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "잘못된 아이디 형식" },
          "500": { "description": "서버 오류" }
        }
      }
    },
    "/api/members": {
      "post": {
        "tags": ["Members"],
        "summary": "회원가입",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberSignupRequest"
              },
              "example": {
                "fullname": "홍길동",
                "username": "honggd01",
                "phone": "010-1234-5678",
                "email": "hong@example.com",
                "password": "pass1234"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "회원가입 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/MemberUser" }
                  }
                }
              }
            }
          },
          "400": { "description": "입력값 검증 실패" },
          "409": { "description": "아이디 중복" },
          "500": { "description": "서버 오류" }
        }
      }
    },
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "주문 목록 조회",
        "responses": {
          "200": {
            "description": "주문 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Order" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "주문 생성",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateRequest"
              },
              "example": {
                "orderNumber": "SANC-2026-0001",
                "userId": 1,
                "totalAmount": 50000,
                "notes": "개별택배"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "주문 생성 성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "주문 단건 조회",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "주문 상세",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" }
              }
            }
          },
          "404": { "description": "주문을 찾을 수 없음" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MemberSignupRequest": {
        "type": "object",
        "required": ["fullname", "username", "phone", "password"],
        "properties": {
          "fullname": { "type": "string", "minLength": 2 },
          "username": {
            "type": "string",
            "pattern": "^[a-z0-9]{4,20}$"
          },
          "phone": { "type": "string", "example": "010-1234-5678" },
          "email": { "type": "string", "format": "email", "nullable": true },
          "password": { "type": "string", "minLength": 4 }
        }
      },
      "MemberUser": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "fullname": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string", "nullable": true },
          "role": { "type": "string", "enum": ["MEMBER", "ADMIN"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "orderNumber": { "type": "string" },
          "status": {
            "type": "string",
            "enum": [
              "PLACED",
              "PRINTING_COMPLETE",
              "WAITING_FOR_SHIPMENT",
              "CANCELLED"
            ]
          },
          "totalAmount": { "type": "number" },
          "notes": { "type": "string", "nullable": true },
          "userId": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "OrderCreateRequest": {
        "type": "object",
        "required": ["orderNumber", "userId"],
        "properties": {
          "orderNumber": { "type": "string" },
          "userId": { "type": "integer" },
          "totalAmount": { "type": "number" },
          "notes": { "type": "string" },
          "status": {
            "type": "string",
            "enum": [
              "PLACED",
              "PRINTING_COMPLETE",
              "WAITING_FOR_SHIPMENT",
              "CANCELLED"
            ]
          }
        }
      }
    }
  }
}